Add set_facility overload for syslog::facility enum
- Add overload to accept syslog::facility enum directly (no cast needed)
- Update test to use enum directly instead of static_cast
- Maintains backward compatibility with int version

+++ net/net-structured_log_stream.test.c++
@@ -573,7 +573,7 @@ auto register_syslogstream_tests()
                         .set_app_name("chained-app")
                         .set_log_level(syslog::severity::warning)
                         .set_sd_id("test-source")
-                        .set_facility(static_cast<int>(syslog::facility::local1))
+                        .set_facility(syslog::facility::local1)
                         .redirect(std::clog);
                 });
             };
